Changes required for running on mobile devices
Removed additional fixture files
Created a new config file for Playwright local
Removed mobile-test file
Added code to the fixture file for patching mobile caps
