+++ package.json
@@ -4,9 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "sample-local-test": "npx playwright test tests/local_test.js",
+    "sample-local-test": "BROWSERSTACK_LOCAL=true npx playwright test tests/local_test.js --config=./playwright-local.config.js",
     "sample-test": "npx playwright test tests/sample_test.js",
-    "mobile-test": "npx playwright test tests/mobile_test.js --project='chrome@Samsung Galaxy S22:13@browserstack'"
+    "mobile-test": "npx playwright test tests/sample_test.js --project='chrome@Samsung Galaxy S22:13@browserstack'"
   },
   "keywords": [],
   "author": "",
